pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentmesh-python"
version = "1.1.0"
description = "Open-source, secure, decentralized agent registration and discovery infrastructure for AI Agent world"
readme = "README.md"
authors = [
    {name = "AgentMesh Team", email = "contact@agentmesh.io"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agent", "discovery", "registry", "a2a", "mcp", "infrastructure"]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
    "prometheus-fastapi-instrumentator>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.23.0",
]
redis = ["redis>=5.0.0"]
postgres = ["asyncpg>=0.29.0"]
grpc = ["grpcio>=1.60.0"]
websocket = ["websockets>=11.0"]
semantic = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0"
]

[project.urls]
Homepage = "https://github.com/agentmesh/agentmesh"
Documentation = "https://agentmesh.io/docs"
Repository = "https://github.com/agentmesh/agentmesh"
Issues = "https://github.com/agentmesh/agentmesh/issues"

[project.scripts]
agentmesh = "agentmesh.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
